+++ pytools4dart/tools/hstools.py
@@ -305,9 +305,9 @@ def _complete_hdr(hdrfile, wavelengths, fwhms):
 
     with open(hdrfile, 'w') as f:
         if wavelengths is not None:
-            hdr = hdr + 'wavelength   = {' + ", ".join([np.str(wvl) for wvl in wavelengths]) + '}'
+            hdr = hdr + '\nwavelength   = {' + ", ".join([np.str(wvl) for wvl in wavelengths]) + '}'
         if fwhms is not None:
-            hdr = hdr + 'fwhm   = {' + ", ".join([np.str(fwhm) for fwhm in fwhms]) + '}'
+            hdr = hdr + '\nfwhm   = {' + ", ".join([np.str(fwhm) for fwhm in fwhms]) + '}'
         f.write(hdr)